backend/pyproject.toml
[tool.poetry]
name = "zest"
version = "0.5.15"
description = ""
authors = ["David B. Adrian <dawidh.adrian@googlemail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.3"
gunicorn = "^23.0.0"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
django-modeltranslation = "^0.19.11"
djangorestframework-simplejwt = "^5.0.0"
dj-rest-auth = "^7.0.0"
django-allauth = {extras = ["socialaccount"], version = "^65.2.0"}
drf-spectacular = ">=0.26.0"
whitenoise = "^6.8.2"
black = "^24.10.0"
coloredlogs = "^15.0.1"
toml = "^0.10.2"
pyyaml = "^6.0.2"
uritemplate = "^4.1.1"
django-auto-prefetching = "^0.2.12"
django-silk = "^5.3.1"
nplusone = "^1.0.0"
django-redis = "^5.4.0"
tqdm = "^4.67.0"
drf-serpy = "^0.5.0"
serpy = "^0.3.1"
psycopg = {extras = ["binary"], version = "^3.2.3"}

[tool.poetry.dev-dependencies]
yapf = "^0.43.0"
mypy = "^1.13"
flake8 = "^7.1.1"
django-stubs = "^5.1.1"
coloredlogs = "^15.0.1"
shiv = "^1.0.8"
coverage = "^7.6"
virtualenv = "^20.27.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
nplusone = "^1.0.0"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
